backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==1.4.56
pydantic==1.10.11
passlib[bcrypt]==1.7.4
python-jose==3.3.0